Design changes for Sykmelding info
Expendable panel for Sykmelding view.
Reorder fields for Sykmelding view.
Add line for separation.
Move expandable info panel to inside og Sykmelding view for employer.
