+++ src/App.less
@@ -1,7 +1,8 @@
 @import '~nav-frontend-core/less/_variabler';
 @import '~nav-frontend-core/less/core.less';
 
-@headerFarge: #d1eff9;
+@headerFarge: #c2eaf7;
+@skillelinjeFarge: #c9c9c9;
 
 @skjermLiten: 30rem;
 @skjermMini: @skjermLiten;
