+++ src/components/Sykmeldingview/Layout/Section/Section.less
@@ -1,12 +1,11 @@
+@import '../../../../App.less';
+
 .opplysning-section {
     margin-bottom: 2rem;
+    border-top: 1px solid @skillelinjeFarge;
+    padding-top: 2rem;
 
     &__title {
         margin-bottom: 1rem;
     }
-}
-
-.opplysning-section:not(:last-of-type) {
-    border-bottom: 1px solid grey;
-    padding-bottom: 2rem;
 }
