+++ src/components/Sykmeldingview/Sections/AktivitetIkkeMuligView.less
@@ -1,0 +1,11 @@
+@import '../../../App.less';
+
+.aktivitet-ikke-mulig {
+    border-top: 1px solid @skillelinjeFarge;
+    padding-top: 2rem;
+    margin-bottom: 2rem;
+
+    &__arsak:not(:last-of-type) {
+        margin-bottom: 2rem;
+    }
+}
