+++ src/components/Sykmeldingview/Sections/MedisinskVurderingView.less
@@ -1,10 +1,11 @@
 @import '../../../App.less';
 
-.periode-view {
-    &__periode {
+.medisinsk-vurdering-view {
+    margin-bottom: 2rem;
+    
+    .diagnoser {
         border-bottom: 1px solid @skillelinjeFarge;
-        padding-bottom: 2rem;
         margin-bottom: 2rem;
+        padding-bottom: 2rem;
     }
 }
-
