+++ src/components/Sykmeldingview/Sections/MedisinskVurderingView.tsx
@@ -2,6 +2,7 @@ import MedisinskVurdering, { AnnenFraverGrunn } from '../../../models/Sykmelding
 import DateFormatter from '../../../utils/DateFormatter';
 import CheckboxEntry from '../Layout/CheckboxEntry/CheckboxEntry';
 import SykmeldingEntry from '../Layout/SykmeldingEntry/SykmeldingEntry';
+import './MedisinskVurderingView.less';
 
 interface MedisinskVurderingViewProps {
     medisinskVurdering?: MedisinskVurdering;
@@ -19,7 +20,7 @@ const MedisinskVurderingView: React.FC<MedisinskVurderingViewProps> = ({ medisin
         }
 
         return (
-            <div style={{ marginBottom: '2rem' }}>
+            <div className="medisinsk-vurdering-view">
                 {!!medisinskVurdering.hovedDiagnose?.tekst && (
                     <SykmeldingEntry title="Diagnose" mainText={medisinskVurdering?.hovedDiagnose?.tekst} sladd />
                 )}
@@ -34,22 +35,18 @@ const MedisinskVurderingView: React.FC<MedisinskVurderingViewProps> = ({ medisin
     }
 
     return (
-        <div style={{ marginBottom: '2rem' }}>
-            {!!medisinskVurdering.hovedDiagnose?.tekst && (
-                <SykmeldingEntry title="Diagnose" mainText={medisinskVurdering?.hovedDiagnose?.tekst} />
-            )}
-            {medisinskVurdering.biDiagnoser.map((bidiagnose, index) => {
-                if (bidiagnose.tekst) {
-                    return <SykmeldingEntry key={index} title="Bidiagnose" mainText={bidiagnose.tekst} />;
-                }
-                return null;
-            })}
-            {!!medisinskVurdering.yrkesskadeDato && (
-                <SykmeldingEntry
-                    title="Skadedato"
-                    mainText={DateFormatter.toReadableDate(medisinskVurdering.yrkesskadeDato)}
-                />
-            )}
+        <div className="medisinsk-vurdering-view">
+            <div className="diagnoser">
+                {!!medisinskVurdering.hovedDiagnose?.tekst && (
+                    <SykmeldingEntry title="Diagnose" mainText={medisinskVurdering?.hovedDiagnose?.tekst} />
+                )}
+                {medisinskVurdering.biDiagnoser.map((bidiagnose, index) => {
+                    if (bidiagnose.tekst) {
+                        return <SykmeldingEntry key={index} title="Bidiagnose" mainText={bidiagnose.tekst} />;
+                    }
+                    return null;
+                })}
+            </div>
             {!!(
                 medisinskVurdering.annenFraversArsak?.grunn && medisinskVurdering.annenFraversArsak?.grunn.length > 0
             ) && (
@@ -70,6 +67,12 @@ const MedisinskVurderingView: React.FC<MedisinskVurderingViewProps> = ({ medisin
                 show={medisinskVurdering.yrkesskade}
                 checkboxText="Sykdommen kan skyldes en yrkesskade/yrkessykdom"
             />
+            {!!medisinskVurdering.yrkesskadeDato && (
+                <SykmeldingEntry
+                    title="Skadedato"
+                    mainText={DateFormatter.toReadableDate(medisinskVurdering.yrkesskadeDato)}
+                />
+            )}
         </div>
     );
 };
