+++ src/components/Sykmeldingview/Sykmeldingview.less
@@ -1,4 +1,16 @@
+@import '../../App.less';
+
 .sykmeldingsview {
-    border-top: 1px solid grey;
+    border-top: 1px solid @skillelinjeFarge;
+    padding-top: 2rem;
+}
+
+.dato-sykmeldingen-ble-skrevet {
+    margin-bottom: 2rem;
+}
+
+.dato-sykmeldingen-ble-skrevet__arbeidsgiver {
+    margin-bottom: 2rem;
+    border-top: 1px solid @skillelinjeFarge;
     padding-top: 2rem;
 }
