+++ src/pages/sykmeldingkvittering/SykmeldingkvitteringPage.tsx
@@ -77,11 +77,11 @@ const SykmeldingkvitteringPage: React.FC = () => {
             </Spacing>
 
             <Spacing>
-                <Sykmeldingsopplysninger sykmelding={sykmelding} arbeidsgiver={false} />
+                <Sykmeldingsopplysninger sykmelding={sykmelding} expandedDefault={false} arbeidsgiver={false} />
             </Spacing>
 
             {sykmelding.sykmeldingStatus.statusEvent === 'SENDT' && (
-                <Sykmeldingsopplysninger sykmelding={sykmelding} arbeidsgiver />
+                <Sykmeldingsopplysninger sykmelding={sykmelding} expandedDefault={false} arbeidsgiver />
             )}
         </PageWrapper>
     );
